SITL: Split 7 series impl to own file
